tuguldur/requirements.txt
bycycle==1.1.0
contourpy==1.2.0
cycler==0.12.1
dcor==0.6
emd==0.6.2
fonttools==4.46.0
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.41.1
matplotlib==3.8.2
neurodsp==2.2.1
numba==0.58.1
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
scipy==1.11.4
six==1.16.0
sparse==0.14.0
tabulate==0.9.0
tzdata==2023.3
